+++ TombEngine/Game/control/trigger.cpp
@@ -282,7 +282,7 @@ void RefreshCamera(short type, short* data)
 			{
 				Camera.number = value;
 
-				if ((Camera.timer < 0) || (Camera.type == CameraType::Look) || (Camera.type == CameraType::Combat))
+				if (Camera.timer < 0 || !TestLockedCamera())
 				{
 					Camera.timer = -1;
 					targetOk = 0;
@@ -297,7 +297,7 @@ void RefreshCamera(short type, short* data)
 			break;
 
 		case TO_TARGET:
-			if (Camera.type == CameraType::Look || Camera.type == CameraType::Combat)
+			if (!TestLockedCamera())
 				break;
 
 			Camera.item = &g_Level.Items[value];
@@ -651,7 +651,7 @@ void TestTriggers(int x, int y, int z, FloorInfo* floor, Activator activator, bo
 
 			Camera.number = value;
 
-			if (Camera.type == CameraType::Look || Camera.type == CameraType::Combat && !(g_Level.Cameras[value].Flags & 3))
+			if (!TestLockedCamera())
 				break;
 
 			if (triggerType == TRIGGER_TYPES::COMBAT)